refactor(React-Training): Created separate components for the line chart and the metrics card
Refactored and cleaned the DashboardPage.tsx file by abstracting sections of it into separate components.

[ICTRAINING-2211]
