+++ src/components/LineChart/LineChart.module.css
@@ -1,0 +1,3 @@
+.container {
+  /* Add styles here */
+}
