+++ src/components/MetricCard/MetricCard.module.css
@@ -1,0 +1,8 @@
+.metricCard {
+  text-align: center;
+}
+
+.metricCard:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
