+++ src/components/MetricCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MetricCard';
